+++ R/cohens_d.R
@@ -188,7 +188,7 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
       out[, colnames(out) %in% c(types, "CI_low", "CI_high")] * correction
   }
 
-  class(out) <- c("effectsize_table", class(out))
+  class(out) <- c("effectsize_table","see_effectsize_table", class(out))
   return(out)
 }