+++ NEWS.md
@@ -3,6 +3,8 @@
 ## New features
 
 - `eta_squared()`, `omega_squared()` and `epsilon_squared()` now works with `gam` models.
+- New function `equivalence_test()` for effect sizes.
+- New plotting methods in the `see` package.
 
 # effectsize 0.3.0